modularised code for badges section
