+++ app/components/profile-components/data.ts
@@ -1,0 +1,410 @@
+import type { Badge, Category } from './types';
+
+// Category definitions with display names and icons
+export const categories: Category[] = [
+  { id: 'code', name: 'Code Mastery', icon: '🧩' },
+  { id: 'bugs', name: 'Bug Hunter', icon: '🐛' },
+  { id: 'languages', name: 'Polyglot', icon: '🌐' },
+  { id: 'mentor', name: 'Mentorship', icon: '🧠' },
+  { id: 'qa', name: 'Quality Assurance', icon: '🔍' },
+  { id: 'creative', name: 'Documentation & Ideas', icon: '💡' },
+  { id: 'points', name: 'Points Progression', icon: '⭐' },
+  { id: 'lang_mastery', name: 'Language Competition', icon: '🏆' },
+  { id: 'mastery', name: 'Ultimate Mastery', icon: '👑' },
+];
+
+export const achievementData: Badge[] = [
+  // CODING PATH (All unlocked, including diamond)
+  {
+    id: 1,
+    title: 'Pull Request Pro',
+    description: 'Merged 1+ Pull Request.',
+    icon: '/icon_badge.png',
+    unlocked: true,
+    tier: 'bronze',
+    date: '2025-05-10',
+    category: 'code',
+    position: 1,
+  },
+  {
+    id: 2,
+    title: 'Merge Master',
+    description: 'Merged 5+ Pull Requests.',
+    icon: '/icon_badge.png',
+    unlocked: true,
+    tier: 'silver',
+    date: '2025-05-12',
+    category: 'code',
+    position: 2,
+    requires: [1],
+  },
+  {
+    id: 3,
+    title: 'Engineer Overachiever',
+    description: 'Merged 10+ Pull Requests.',
+    icon: '/icon_badge.png',
+    unlocked: true,
+    tier: 'gold',
+    date: '2025-05-14',
+    category: 'code',
+    position: 3,
+    requires: [2],
+  },
+  {
+    id: 4,
+    title: 'Code Crusader',
+    description: 'Merged 20 Pull Requests.',
+    icon: '/icon_badge.png',
+    unlocked: true,
+    tier: 'diamond',
+    date: '2025-05-15',
+    category: 'code',
+    position: 4,
+    requires: [3],
+  },
+
+  // BUG HUNTING PATH (Unlocked up to gold)
+  {
+    id: 5,
+    title: 'Bug Spotter',
+    description: 'Got 1 bug accepted (issue).',
+    icon: '/icon_badge.png',
+    unlocked: true,
+    tier: 'bronze',
+    date: '2025-05-08',
+    category: 'bugs',
+    position: 1,
+  },
+  {
+    id: 6,
+    title: 'Bug Hunter',
+    description: 'Got 5 bugs accepted (issue).',
+    icon: '/icon_badge.png',
+    unlocked: true,
+    tier: 'silver',
+    date: '2025-05-12',
+    category: 'bugs',
+    position: 2,
+    requires: [5],
+  },
+  {
+    id: 7,
+    title: 'Bug Exterminator',
+    description: 'Got 10 bugs accepted (issue).',
+    icon: '/icon_badge.png',
+    unlocked: true,
+    tier: 'gold',
+    date: '2025-05-14',
+    category: 'bugs',
+    position: 3,
+    requires: [6],
+  },
+
+  // POLYGLOT PATH (Only silver unlocked)
+  {
+    id: 8,
+    title: 'Polyglot',
+    description: 'Contributed in 3 different languages.',
+    icon: '/icon_badge.png',
+    unlocked: true,
+    tier: 'silver',
+    date: '2025-05-11',
+    category: 'languages',
+    position: 1,
+  },
+  {
+    id: 9,
+    title: 'Jack of All Trades',
+    description: 'Contributed in 5 different languages.',
+    icon: '/icon_badge.png',
+    unlocked: false,
+    tier: 'gold',
+    progress: 40,
+    category: 'languages',
+    position: 2,
+    requires: [8],
+  },
+
+  // MENTORSHIP PATH (Only bronze unlocked)
+  {
+    id: 10,
+    title: 'The Scholar',
+    description: 'Help out 1 person with their issue solving (in group).',
+    icon: '/icon_badge.png',
+    unlocked: true,
+    tier: 'bronze',
+    date: '2025-05-09',
+    category: 'mentor',
+    position: 1,
+  },
+  {
+    id: 11,
+    title: 'The Facilitator',
+    description: 'Help out 3 people with issue solving (in group).',
+    icon: '/icon_badge.png',
+    unlocked: false,
+    tier: 'silver',
+    progress: 30,
+    category: 'mentor',
+    position: 2,
+    requires: [10],
+  },
+  {
+    id: 12,
+    title: 'The Oracle',
+    description: 'Help out 5 people with issue solving (in group).',
+    icon: '/icon_badge.png',
+    unlocked: false,
+    tier: 'gold',
+    progress: 10,
+    category: 'mentor',
+    position: 3,
+    requires: [11],
+  },
+
+  // QUALITY ASSURANCE PATH (Unlocked up to gold)
+  {
+    id: 13,
+    title: 'Issue Assistant',
+    description: 'Got 1 testing related PR merged.',
+    icon: '/icon_badge.png',
+    unlocked: true,
+    tier: 'bronze',
+    date: '2025-05-07',
+    category: 'qa',
+    position: 1,
+  },
+  {
+    id: 14,
+    title: 'Quality Assurer',
+    description: 'Got 5 testing related PRs merged.',
+    icon: '/icon_badge.png',
+    unlocked: true,
+    tier: 'silver',
+    date: '2025-05-10',
+    category: 'qa',
+    position: 2,
+    requires: [13],
+  },
+  {
+    id: 15,
+    title: 'Full-fledged Alchemist',
+    description: 'Contribute 10 testing related PRs.',
+    icon: '/icon_badge.png',
+    unlocked: true,
+    tier: 'gold',
+    date: '2025-05-12',
+    category: 'qa',
+    position: 3,
+    requires: [14],
+  },
+
+  // CREATIVITY PATH (Only bronze unlocked)
+  {
+    id: 17,
+    title: 'Doc Shaman',
+    description: 'Get 2 documentation PRs accepted.',
+    icon: '/icon_badge.png',
+    unlocked: true,
+    tier: 'bronze',
+    date: '2025-05-06',
+    category: 'creative',
+    position: 1,
+  },
+  {
+    id: 16,
+    title: 'Highly Creative',
+    description: 'Get 2 feature suggestions accepted.',
+    icon: '/icon_badge.png',
+    unlocked: false,
+    tier: 'silver',
+    progress: 10,
+    category: 'creative',
+    position: 2,
+  },
+
+  // POINTS PATH (All unlocked, including diamond)
+  {
+    id: 18,
+    title: 'Shawn',
+    description: 'Reach 250 bounty points.',
+    icon: '/icon_badge.png',
+    unlocked: true,
+    tier: 'bronze',
+    date: '2025-05-06',
+    category: 'points',
+    position: 1,
+  },
+  {
+    id: 19,
+    title: 'Ricochet',
+    description: 'Reach 500 bounty points.',
+    icon: '/icon_badge.png',
+    unlocked: true,
+    tier: 'silver',
+    date: '2025-05-08',
+    category: 'points',
+    position: 2,
+    requires: [18],
+  },
+  {
+    id: 20,
+    title: 'Firefinch',
+    description: 'Reach 750 bounty points.',
+    icon: '/icon_badge.png',
+    unlocked: true,
+    tier: 'gold',
+    date: '2025-05-10',
+    category: 'points',
+    position: 3,
+    requires: [19],
+  },
+  {
+    id: 21,
+    title: 'The Scythe of Vyse',
+    description: 'Reach 1000 bounty points.',
+    icon: '/icon_badge.png',
+    unlocked: true,
+    tier: 'diamond',
+    date: '2025-05-12',
+    category: 'points',
+    position: 4,
+    requires: [20],
+  },
+
+  // LANGUAGE MASTERY PATHS (Mixed progress across all 5 languages)
+  // Rust (Only silver unlocked)
+  {
+    id: 23,
+    title: 'Crabby Coders',
+    description: 'Most Rust issues solved (2nd place).',
+    icon: '/icon_badge.png',
+    unlocked: true,
+    tier: 'silver',
+    date: '2025-05-05',
+    category: 'lang_mastery',
+    position: 1,
+  },
+  {
+    id: 22,
+    title: 'Pioneers of Maximum Efforts',
+    description: 'Most Rust issues solved (1st place).',
+    icon: '/icon_badge.png',
+    unlocked: false,
+    tier: 'gold',
+    progress: 40,
+    category: 'lang_mastery',
+    position: 2,
+    requires: [23],
+  },
+  // Zig (Up to gold)
+  {
+    id: 24,
+    title: "Salamander's Toboggan",
+    description: 'Most Zig issues solved (2nd place).',
+    icon: '/icon_badge.png',
+    unlocked: true,
+    tier: 'silver',
+    date: '2025-05-04',
+    category: 'lang_mastery',
+    position: 3,
+  },
+  {
+    id: 25,
+    title: "Strimlander's Spirit",
+    description: 'Most Zig issues solved (1st place).',
+    icon: '/icon_badge.png',
+    unlocked: true,
+    tier: 'gold',
+    date: '2025-05-05',
+    category: 'lang_mastery',
+    position: 4,
+    requires: [24],
+  },
+  // Python (Only silver unlocked)
+  {
+    id: 27,
+    title: 'Basilisk Defanged',
+    description: 'Most Python issues solved (2nd place).',
+    icon: '/icon_badge.png',
+    unlocked: true,
+    tier: 'silver',
+    date: '2025-05-03',
+    category: 'lang_mastery',
+    position: 5,
+  },
+  {
+    id: 26,
+    title: 'Mambo Montalvo',
+    description: 'Most Python issues solved (1st place).',
+    icon: '/icon_badge.png',
+    unlocked: false,
+    tier: 'gold',
+    progress: 50,
+    category: 'lang_mastery',
+    position: 6,
+    requires: [27],
+  },
+  // Go (Only silver unlocked)
+  {
+    id: 29,
+    title: 'Primal Suricats',
+    description: 'Most Go issues solved (2nd place).',
+    icon: '/icon_badge.png',
+    unlocked: true,
+    tier: 'silver',
+    date: '2025-05-03',
+    category: 'lang_mastery',
+    position: 7,
+  },
+  {
+    id: 28,
+    title: 'Apex Gophers',
+    description: 'Most Go issues solved (1st place).',
+    icon: '/icon_badge.png',
+    unlocked: false,
+    tier: 'gold',
+    progress: 20,
+    category: 'lang_mastery',
+    position: 8,
+    requires: [29],
+  },
+  // JS/TS (Up to gold)
+  {
+    id: 31,
+    title: 'Trop Killers',
+    description: 'Most JS/TS issues solved (2nd place).',
+    icon: '/icon_badge.png',
+    unlocked: true,
+    tier: 'silver',
+    date: '2025-05-03',
+    category: 'lang_mastery',
+    position: 9,
+  },
+  {
+    id: 30,
+    title: 'Forge Dwellers',
+    description: 'Most JS/TS issues solved (1st place).',
+    icon: '/icon_badge.png',
+    unlocked: true,
+    tier: 'gold',
+    date: '2025-05-04',
+    category: 'lang_mastery',
+    position: 10,
+    requires: [31],
+  },
+
+  // MASTERY PATH (None unlocked, cross-category)
+  {
+    id: 32,
+    title: 'Weekly Warrior',
+    description: 'Make a submission accepted for each weekly challenge.',
+    icon: '/icon_badge.png',
+    unlocked: false,
+    tier: 'gold',
+    progress: 90,
+    category: 'mastery',
+    position: 1,
+    requires: [2, 5, 8],
+  },
+];
